requirements.txt
streamlit
langchain
langchain_chroma
langchain_community
python-dotenv
langchain_groq
langchain_huggingface
langchain_chroma
langchain_core
pypdf
wikipedia
python-dotenv
faiss-cpu
duckduckgo-search
SQLAlchemy
mysql-connector-python
pytube
unstructured
numexpr
cassio
openai
tiktoken
PyPDF2
gradio
langgraph
langchain-openai
ollama
langchain-ollama
langchain-deepseek-official